Put all the slides in history. Remove slides.
Slides were stored in dictionary and referenced in the history
array. This change removes the indirection. Now history holds the slides
directly. The indirection wasn't helping because the editing and
displaying model variables still referenced the history array and had to
be updated at the end of every history reordering. The resulting code is
shorter and easier to understand.